Adapt patches nose plugin to pytest
